+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ## 0.12.0
 ### Features
-* new macros `list*` and `begin*` that run their arguments in parallel if they async
+* new macro `begin*` that run their arguments in parallel if they async
+* add typecheck to rest of the functions
+* new `real?` and `pluck` functions
+### Bug fixes
+* fix `find` and `filter` to check if function from argument return nil
+* fix accessing FileReader object (try to copy read only prototype)
 
 ## 0.11.2
 ### Breaking